move custom.Content to cmd.Custom for greater reusability
